+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # NEXT RELEASE
 * Add filters to `getCurrentState()` return value
+* Remove diacritics from text when filtering for hashtag matches. Fixes bug where german/french etc. tags weren't filtering correctly bc tag names have already had diacritics stripped
 
 ### v0.54.1
 * (Demo App) Memoization optimizations for `StatusComponent` props, `TrendingInfo` sections (the most expensive renders according to profiler)